Add a test to verify it won't clobber present files.
If a redirect_from entry targets a content location which still exists,
the redirect will not be created because of the directory existence
protection in lib/nanoc-redirector.rb:20. However, this implentation
also prevents the creation of redirects from directories that have other
content. Such as an `articles` directory. This test verifies the
essential part of the current behavior so that future commits to allow
redirects from directories with content can be created.