Add a failing test for redirects from directories with content.
Due to the conditional in lib/nanoc-redirector.rb:20 it's not possible
to create redirects for directories that have existing content. This is
good when that content is an `index.html` file which would be
overwritten, but we should be able to redirect from the index of a
directory that only exists with nested content. This failing test and
new fixture demonstrates what we want to achieve.